requirements-vendor.txt
Django==1.9
django-queryset-csv
GoogleAppEngineCloudStorageClient
google-api-python-client